requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==2.2.1
django-environ==0.4.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
environ==1.0
Faker==0.9.1
freezegun==0.3.15
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.2.0
openapi-codec==1.3.2
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.11.2
